Use canned API responses for development
Spin up a dummy server with "yarn serve" serving canned API responses
that exercise the code base, and use these responses for development via
a new "env" module for environment settings

Typicode's JSON Server compared favourably next to two other servers I
assessed. It has a lot of neat functionality that both is of no use at
all and does not work for the GW2 API but provides immediate access to
an Express 4.x server, which makes it easy to emulate the API. It would
probably be easy to run Express directly, but JSON Server comes with
some nice-to-have middleware.