Show achievement requirement on hover
This is not super helpful, and may in fact be more distracting. It's
useful for new or very rare achievements to learn their requirements,
such as with the "Daily Demolisher" introduced in the 2018-07-10 balance
patch, but most achievements that could actually benefit from a proper
requirement description -- for instance all fractals dailies -- have
worthless requirement descriptions. Also, the "Daily Demolisher"
achievement's requirement wrongly listed it as a Vabbi achievement when
in fact it was the "Maws of Torment" Desolation meta event, so yeah.